Améloaration

+++ dictionnaire.py
@@ -7,13 +7,16 @@ def recupere_donnes():
     """
     Retourne les données sous forme d'une liste de dictionnaires
     """
-    local_file = "college_finistere.csv" #nom donné au fichier
-    # recuperation des données depuis l'url
-    url = "https://geobretagne.fr/geoserver/cd29/wfs?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0&TYPENAMES=cd29%3Acolleges_29&OUTPUTFORMAT=csv"
-    data = requests.get(url).content # on récupère les données sous forme de bytes
-    # création d'un fichier et écriture des données
-    with open(local_file, 'wb') as csvfile: # 'w' pour write et 'b' pour bytes
-        csvfile.write(data)
+    local_file = "college_finistere.csv"
+
+    if not os.path.exists(local_file):
+        # recuperation des données depuis l'url
+        url = "https://geobretagne.fr/geoserver/cd29/wfs?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0&TYPENAMES=cd29%3Acolleges_29&OUTPUTFORMAT=csv"
+        data = requests.get(url).content # on récupère les données sous forme de bytes
+        print("telecarger")
+        # création d'un fichier et écriture des données
+        with open(local_file, 'wb') as csvfile: # 'w' pour write et 'b' pour bytes
+            csvfile.write(data)
     
     donnes = []
     # ouverture du fichier précédemment créé
@@ -26,3 +29,4 @@ def recupere_donnes():
 etablissements = recupere_donnes()
 
 print(etablissements[0:10])
+
